+++ configure.py
@@ -513,6 +513,7 @@ for name in ['build',
              'metrics',
              'parser',
              'state',
+             'status',
              'string_piece_util',
              'util',
              'version']:
@@ -579,6 +580,7 @@ for name in ['build_log_test',
              'manifest_parser_test',
              'ninja_test',
              'state_test',
+             'status_test',
              'string_piece_util_test',
              'subprocess_test',
              'test',
